added tons of random shit
