--- server/tests/routes/map/map.test.js
@@ -1,100 +1,0 @@
-const request = require("supertest");
-
-// Test the map routes
-describe("testing the map routse", () => {
-  describe("GET /map", () => {
-    describe("when user is authenticated", () => {
-      test("with an ID of a map, it should return status code 201", async () => {
-        // Test logic for a request with an ID of a map
-      });
-
-      test("with no ID of a map, it should return status code 201", async () => {
-        // Test logic for a request without an ID of a map
-      });
-    });
-
-    describe("when user is not authenticated", () => {
-      it("should return status code 401", async () => {
-        // Test logic for an unauthenticated user
-      });
-    });
-  });
-
-  describe("GET /map/:id", () => {
-    describe("when user is authenticated", () => {
-      test("given an ID of a map, it should return status code 201", async () => {
-        // Test logic for a valid ID of a map
-      });
-
-      test("given an invalid ID of a map, it should return status code 404", async () => {
-        // Test logic for an invalid ID of a map
-      });
-    });
-
-    describe("when user is not authenticated", () => {
-      it("should return status code 401", async () => {
-        // Test logic for an unauthenticated user
-      });
-    });
-  });
-
-  describe("PUT /map", () => {
-    describe("when user is authenticated", () => {
-      test("given map file details, it should return status code 201", async () => {
-        // Test logic for map file details provided
-      });
-
-      test("given no map file details, it should return status code 400", async () => {
-        // Test logic for no map file details provided
-      });
-    });
-
-    describe("when user is not authenticated", () => {
-      it("should return status code 401", async () => {
-        // Test logic for an unauthenticated user
-      });
-    });
-  });
-
-  describe("POST /map/:id", () => {
-    describe("when user is authenticated", () => {
-      test("when ID is not valid for the user to update, it should return status code 400", async () => {
-        // Test logic for an invalid ID for the user to update
-      });
-
-      test("when contents of the request are not correct, it should return status code 400", async () => {
-        // Test logic for incorrect request contents
-      });
-
-      test("when ID is valid and all contents are valid, it should return status code 200", async () => {
-        // Test logic for a valid ID and contents
-      });
-    });
-
-    describe("when user is not authenticated", () => {
-      it("should return status code 401", async () => {
-        // Test logic for an unauthenticated user
-      });
-    });
-  });
-
-  describe("POST /map/duplicate", () => {
-    describe("when user is authenticated", () => {
-      test("with a valid map ID, it should successfully duplicate the map and return 201", async () => {
-        // Test logic for duplicating a map with a valid ID
-      });
-
-      test("with an invalid map ID, it should return a specific error and return 400", async () => {
-        // Test logic for duplicating a map with an invalid ID
-      });
-    });
-
-    describe("when user is not authenticated", () => {
-      it("should return status code 401", async () => {
-        // Test logic for an unauthenticated user
-      });
-    });
-  });
-});
-
-
