+++ server/tests/routes/map/map.test.ts
