+++ server/tests/routes/map/mapController.test.ts
@@ -1,20 +1,13 @@
 import Fixtures from "node-mongodb-fixtures";
 import mongoose from "mongoose";
-import { mapSchema } from "./models/Map";
-import map1 from "./models/tests/map1.json";
+import { mapSchema } from "../../models/Map";
+import map1 from "../../models/tests/map1.json";
 
 describe("Testing connection to Mongodb server", () => {
   it("it should exit gracefully", async () => {
-    const fixtures = new Fixtures({
-      dir: "tests/fixtures",
-      mute: true, // do not mute the log output
-    });
+    const mockdb = await mongoose.connect("mongodb://127.0.0.1:27017");
 
-    await fixtures.connect("mongodb://localhost:27017/");
-
-    await mongoose.connect("mongodb://127.0.0.1:27017");
-
-    await fixtures.disconnect();
+    await mockdb.disconnect();
   });
 
   test("if unload and load worked by finding a Person", async () => {
@@ -27,9 +20,10 @@ describe("Testing connection to Mongodb server", () => {
       .connect("mongodb://localhost:27017/")
       .then(() => fixtures.unload())
       .then(() => fixtures.load())
-      .catch((e) => console.error(e));
+      .catch((e) => console.error(e))
+      .then(() => fixtures.disconnect());
 
-    await mongoose.connect("mongodb://127.0.0.1:27017");
+    const mockdb = await mongoose.connect("mongodb://127.0.0.1:27017");
 
     const personSchema = new mongoose.Schema({
       name: String,
@@ -42,28 +36,22 @@ describe("Testing connection to Mongodb server", () => {
 
     console.log(paul[0].name);
 
-    await fixtures.disconnect();
+    await mockdb.disconnect();
   });
 });
 
 describe("Crud Map operations", () => {
   test("if CREATE a map works", async () => {
-    const fixtures = new Fixtures({
-      dir: "tests/fixtures",
-      mute: true, // do not mute the log output
-    });
-    await fixtures.connect("mongodb://localhost:27017/");
-
-    await mongoose.connect("mongodb://127.0.0.1:27017");
+    const mockdb = await mongoose.connect("mongodb://127.0.0.1:27017");
 
     const Map = mongoose.model("map", mapSchema);
 
     const map = new Map(map1);
     const result = await map.save();
 
-    console.log(result.mapName)
+    console.log(result.mapName);
 
-    await fixtures.disconnect();
+    await mockdb.disconnect();
   });
 
   test("if READ a map works", async () => {
@@ -72,14 +60,16 @@ describe("Crud Map operations", () => {
       mute: true, // do not mute the log output
       filter: "map.*",
     });
-    await fixtures.connect("mongodb://localhost:27017/");
-    await fixtures.unload();
-    await fixtures.load();
+    await fixtures
+      .connect("mongodb://localhost:27017/")
+      .then(() => fixtures.unload())
+      .then(() => fixtures.load())
+      .catch((e) => console.error(e))
+      .then(() => fixtures.disconnect());
 
-    await mongoose.connect("mongodb://127.0.0.1:27017");
+    const mockdb = await mongoose.connect("mongodb://127.0.0.1:27017");
 
     const Map = mongoose.model("map", mapSchema);
-
 
     const retrievedmap = await Map.find({ mapName: "Sample Map 1" });
 
@@ -88,14 +78,10 @@ describe("Crud Map operations", () => {
 
     console.log(retrievedmap[0]);
 
-    await fixtures.disconnect();
+    await mockdb.disconnect();
   });
 
-  test("if UPDATE a map works", async () => {
+  test("if UPDATE a map works", async () => {});
 
-  })
-
-  test("if DELETE a map works", async () => {
-
-  })
+  test("if DELETE a map works", async () => {});
 });
